WIP feedback changes
Bigger font across the board - everything, including titles, table of contents, inside the drawers and the subtitle text on the slides. Relative to screen size. There are some screens where the font is almost illegible.
- Increased the article content, title and subtitle of slides, and table of contents
- Done

'Parrallax' background effects for each slide. To confirm: the background effect should only be on 5 slides:  A beekeeper..., Emerging leaders..., Data drives student success..., A multisectoral approach..., University degree program...// TBD

Change the Afghanistan story photo for slide 1 (with the drawer photo), and add a slide 2
- I'm not sure if I'm tracking on this. Could you elaborate on the section, the drawer photo and slide 2? Is the current build missing a section?

Animating numbers counting up for certain slide statistics,

'Share' icon in the top navigation bar

Table of Contents circle animation

Right now when you click on a profile story in the menu, it takes you to the second slide for that story. It should go to the first slide (ie "Data drives student success" should take you to the slide with Jennifer's picture, not the silos).
- Done

Add link to "CEO Message" in the top level nav bar. Add a little space between all of these links.
- Done

Photography credits, both in bottom bar and left drawer.

On the slides: move the photo credit link to the left side, and add small navigation arrows on the right side, near where the scroll bar would appear. A small up arrow in the upper right corner, which on click would take you to the previous slide, and a small down arrow in the bottom right.
- Does this mean that the Slides should move left/right? This wasn't in Searching for Syria, so I want to be certain.

When you open a drawer, click anywhere off the drawer to close and return to the slide (right now you have to click directly on the "x")
- Done
Make “Read Story” and “Learn More” blue buttons more prominent so people know to click on them. Possibly have them slide/fade in like the title/subtitle text and pulse once. Make them a tad bigger.
- Done
The overlay on the Obey Assery/Tanzania slide is too dark. Same with the photo of the grain silos with the quote from Jennifer Carlson.
- Done
On Supporting Myanmar, could the 3 check mark lines come in separately, and/or draw the check marks in.

In the drawer for Patrick’s letter – his photo is very small. It looks like it’s the same size as the thumbnail on the slide. Can we make that larger? It does not need to be square.
- Done
Yatin photo needs to not be downsized from the top (“Emerging leaders…” story). It cuts off his head when the browser window gets wider. If that can not be prevented, we will have to swap the photo.
- Done
Add drawer photo for “Data drives student success in America’s heartland” story
- Done
On some of the drawers the box shadow is creeping into the bottom left corner of the expanded drawer.
- Done

Change Copyright year at the bottom to 2018, not 2017
- Done

Add anchor tags to each of the 10 chapters, so that these stories can be promoted individually on social media. (i.e. /annual-report-2017/#beekeeper)
- I have sections numbered (e.g. section1, section2, can you send a list of section names?

Signed-off-by: Sean Harrington <sean@never-not.com>